requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.0.200
groq>=0.3.0
serpapi>=0.1.0